Add list of chess players with capitalization button
When clicking the button the capitalization changes based on the count variable.
Count starts at 0. If count is even (as in the beginning situation), the player names are capitalized.
If the count is odd, the player names are displayed in lower case again.
There probably are better ways to implement this alternating behavior but this is just a proof of concept.