Add Cypress tests for list of players
In order to run the tests, you need a local server running.
Since at the moment the app is just a single index.html file, the http-server package is used to serve that file.
At the moment the baseUrl is hardcoded to point to the default address used by http-server.
Also in the Cypress configuration fixtures and video have been disabled for now.